+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @web3inbox/core
 
+## 0.0.7
+
+### Patch Changes
+
+- Add checks on account
+
 ## 0.0.6
 
 ### Patch Changes